resolving connection string and schema name
